pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "geo-sim"
version = "0.1.0"
description = "Geo-Sim: Geopolitical Intelligence Platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "neo4j>=5.20.0",
    "redis>=5.0.4",
    "psycopg2-binary>=2.9.9",
    "qdrant-client>=1.9.1",
    "confluent-kafka>=2.4.0",
    "openai>=1.30.1",
    "anthropic>=0.28.0",
    "spacy>=3.7.4",
    "flair>=0.13.1",
    "numpy>=1.26.4",
    "feedparser>=6.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "black>=24.4.2",
    "isort>=5.13.2",
    "mypy>=1.10.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["ingestion", "state", "knowledge_graph", "rules", "world_model", "learning", "simulator", "api", "shared"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
